entries going into html through pyramid template, moved code to proper location
