+++ cnxarchive/templates/entries.html
@@ -15,7 +15,7 @@
             {{module.authors}}
       </i> </b>
       <br></br>
-         {{module.abstract}}
+         {{module.abstract | urlencode}}
       <br></br>
       <img src="http://i.creativecommons.org/l/by/4.0/80x15.png" />
       <a href="http://creativecommons.org/licenses/by/4.0/">
